BACKUP/web_22h45_30.3.2026/TutorAI_Bio-Tree/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
beautifulsoup4==4.13.4
blinker==1.9.0
bottle==0.13.4
cachelib==0.13.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
clr_loader==0.2.7.post0
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.5
cssselect==1.3.0
cycler==0.12.1
docx2python==3.5.0
filelock==3.18.0
Flask==3.1.0
Flask-Caching==2.3.1
Flask-HTTPAuth==4.8.0
fonttools==4.58.4
gcloud==0.18.3
gdown==5.2.0
git-filter-repo==2.47.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.71.2
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jwcrypto==1.5.6
kiwisolver==1.4.8
lxml==6.0.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
paragraphs==1.0.1
pefile==2023.2.7
pillow==10.4.0
proto-plus==1.26.1
protobuf==5.29.5
proxy_tools==0.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
pymdown-extensions==10.16
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6-WebEngine==6.9.0
PyQt6-WebEngine-Qt6==6.9.1
PyQt6_sip==13.10.2
Pyrebase4==4.8.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-jwt==4.1.0
pythonnet==3.0.5
pywebview==5.4
pywin32-ctypes==0.2.3
PyYAML==6.0.2
rarfile==4.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rsa==4.9.1
six==1.17.0
soupsieve==2.7
tkcode==0.1.6
tkhtmlview==0.3.1
tqdm==4.67.1
types-html5lib==1.1.11.20250708
types-lxml==2025.3.30
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==1.26.20
Werkzeug==3.1.3
